fix(ffmpeg): MKV raw FFI overhaul — module split, merge, thumbnails, B-frame fix
- Remove CLI fallback for MKV thumbnail embedding
- Remove dead MKV attachment helper functions
- Enable trace logging when -v flag is passed
- Prevent DEBUG logs from appearing without -v flag
- Split monolithic ffmpeg.rs into module directory (remux, merge,
  transcode, thumbnail, metadata, probe)
- Add merge_mkv_raw_ffi() for MKV merge with full stream property
  copying (avg_frame_rate, r_frame_rate, time_base, sample_aspect_ratio)
- Use native Matroska attachments for MKV thumbnail embedding instead
  of attached_pic video streams
- Guard all raw FFI packet loops against AV_NOPTS_VALUE before timestamp
  rescaling — MKV demuxer leaves DTS unset for B-frame content, and
  av_rescale_q_rnd(INT64_MIN) overflows to garbage causing silent packet
  drops and zero duration
