+++ Cargo.toml
@@ -58,7 +58,7 @@ dialoguer = "0.12.0"
 log = { version = "0.4", features = ["kv"] }
 env_logger = { version = "0.11", features = ["kv"] }
 tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter", "fmt", "tracing-log"] }
+tracing-subscriber = { version = "0.3.22", features = ["env-filter", "fmt", "tracing-log"] }
 
 # Configuration
 toml = "0.9.11"
