+++ crates/rdlp-ffmpeg/src/ffmpeg/ffi_helpers.rs
@@ -1,0 +1,137 @@
+//! Unsafe FFI helper functions.
+//!
+//! These encapsulate all `unsafe` FFI operations that lack safe wrappers
+//! in `ffmpeg-the-third`, providing safe call-site signatures. The `unsafe`
+//! blocks are limited to these well-documented helpers.
+
+use crate::error::Result;
+
+use super::FFmpegRunner;
+
+impl FFmpegRunner {
+    /// Reset the codec tag to 0 for container compatibility.
+    ///
+    /// When remuxing between containers, the source codec tag may not be valid
+    /// in the target container. Setting it to 0 lets FFmpeg auto-select.
+    pub(crate) fn clear_codec_tag(params_ptr: *const ffmpeg_the_third::ffi::AVCodecParameters) {
+        // SAFETY: `params_ptr` points to a valid AVCodecParameters allocated by FFmpeg.
+        // Setting codec_tag to 0 is always valid — it tells FFmpeg to auto-select.
+        unsafe {
+            (*(params_ptr as *mut ffmpeg_the_third::ffi::AVCodecParameters)).codec_tag = 0;
+        }
+    }
+
+    /// Copy encoder parameters back to an output stream.
+    ///
+    /// After opening an encoder, its parameters (codec, dimensions, sample rate,
+    /// etc.) must be copied to the corresponding output stream before writing
+    /// the header.
+    pub(crate) fn copy_encoder_params_to_stream(
+        octx: &mut ffmpeg_the_third::format::context::Output,
+        stream_index: usize,
+        encoder_ptr: *const ffmpeg_the_third::ffi::AVCodecContext,
+    ) {
+        // SAFETY: `octx` owns the output context with a valid stream array.
+        // `stream_index` was obtained from a stream added to this context.
+        // `encoder_ptr` points to a valid, opened encoder context.
+        unsafe {
+            let stream_ptr = *(*octx.as_mut_ptr()).streams.add(stream_index);
+            ffmpeg_the_third::ffi::avcodec_parameters_from_context(
+                (*stream_ptr).codecpar,
+                encoder_ptr,
+            );
+        }
+    }
+
+    /// Set the default channel layout for the given number of channels.
+    pub(crate) fn set_default_channel_layout(
+        encoder_ptr: *mut ffmpeg_the_third::ffi::AVCodecContext,
+        channels: i32,
+    ) {
+        // SAFETY: `encoder_ptr` is a valid, pre-open encoder context.
+        // `av_channel_layout_default` populates the ch_layout field in-place.
+        unsafe {
+            ffmpeg_the_third::ffi::av_channel_layout_default(
+                &mut (*encoder_ptr).ch_layout,
+                channels,
+            );
+        }
+    }
+
+    /// Enable VBR (variable bitrate) quality mode on an encoder.
+    pub(crate) fn set_vbr_quality(
+        encoder_ptr: *mut ffmpeg_the_third::ffi::AVCodecContext,
+        quality: i32,
+    ) {
+        // SAFETY: `encoder_ptr` is a valid, pre-open encoder context.
+        // Setting QSCALE flag + global_quality is the standard way to enable VBR.
+        unsafe {
+            (*encoder_ptr).flags |= ffmpeg_the_third::ffi::AV_CODEC_FLAG_QSCALE as i32;
+            (*encoder_ptr).global_quality = quality * ffmpeg_the_third::ffi::FF_QP2LAMBDA;
+        }
+    }
+
+    /// Set the global header flag on an encoder.
+    ///
+    /// Required when the output format needs codec parameters in the container
+    /// header rather than in each packet (e.g., MP4, MKV).
+    pub(crate) fn set_global_header_flag(encoder_ptr: *mut ffmpeg_the_third::ffi::AVCodecContext) {
+        // SAFETY: `encoder_ptr` is a valid, pre-open encoder context.
+        // This flag is required by certain container formats.
+        unsafe {
+            (*encoder_ptr).flags |= ffmpeg_the_third::ffi::AV_CODEC_FLAG_GLOBAL_HEADER as i32;
+        }
+    }
+
+    /// Write thumbnail packets from the thumbnail input to the output context.
+    pub(crate) fn write_thumbnail_packets(
+        thumb_ictx: &mut ffmpeg_the_third::format::context::Input,
+        octx: &mut ffmpeg_the_third::format::context::Output,
+        thumb_ist_index: usize,
+        thumb_ist_time_base: ffmpeg_the_third::Rational,
+        thumb_ost_index: usize,
+    ) -> Result<()> {
+        use crate::error::PostProcessError;
+
+        let thumb_ost_time_base = octx
+            .stream(thumb_ost_index)
+            .ok_or_else(|| {
+                PostProcessError::ffmpeg_failed(format!(
+                    "thumbnail output stream {thumb_ost_index} not found"
+                ))
+            })?
+            .time_base();
+        for result in thumb_ictx.packets() {
+            let (stream, mut packet) =
+                result.map_err(|e| PostProcessError::FFmpegLibraryError {
+                    message: format!("failed to read thumbnail packet: {e}"),
+                })?;
+            if stream.index() == thumb_ist_index {
+                packet.rescale_ts(thumb_ist_time_base, thumb_ost_time_base);
+                packet.set_position(-1);
+                packet.set_stream(thumb_ost_index);
+                packet.write_interleaved(octx).map_err(|e| {
+                    PostProcessError::FFmpegLibraryError {
+                        message: format!("failed to write thumbnail packet: {e}"),
+                    }
+                })?;
+            }
+        }
+        Ok(())
+    }
+
+    /// Configure a stream with `ATTACHED_PIC` disposition (for cover art).
+    ///
+    /// Sets the stream disposition and clears the codec tag. Used for MP4,
+    /// FLAC, OGG, and other containers that embed cover art as a video stream
+    /// with special disposition.
+    pub(crate) fn set_attached_pic_disposition(stream_ptr: *mut ffmpeg_the_third::ffi::AVStream) {
+        // SAFETY: `stream_ptr` is a valid output stream pointer from a live
+        // output context. Setting disposition and clearing codec_tag configures
+        // the stream as cover art.
+        unsafe {
+            (*stream_ptr).disposition = ffmpeg_the_third::ffi::AV_DISPOSITION_ATTACHED_PIC;
+            (*((*stream_ptr).codecpar)).codec_tag = 0;
+        }
+    }
+}
